[BE] refactor: payload 패키지로 분리
