--- src/main/java/com/mini/buting/api/chat/domain/Image.java
@@ -1,4 +1,0 @@
-package com.mini.buting.api.chat.domain;
-
-public class Image {
-}
