--- src/main/java/com/mini/buting/api/chat/domain/Payload.java
@@ -1,3 +1,0 @@
-package com.mini.buting.api.chat.domain;
-
-public sealed interface Payload permits TextPayload, ImagePayload, NoticePayload, VotePayload {}
