+++ src/main/java/com/mini/buting/api/chat/domain/payload/Image.java
@@ -1,0 +1,4 @@
+package com.mini.buting.api.chat.domain.payload;
+
+public class Image {
+}
