+++ src/main/java/com/mini/buting/api/chat/domain/payload/NoticePayload.java
@@ -1,4 +1,4 @@
-package com.mini.buting.api.chat.domain;
+package com.mini.buting.api.chat.domain.payload;
 
 import java.time.LocalDateTime;
