+++ src/main/java/com/mini/buting/api/chat/domain/payload/TextPayload.java
@@ -1,4 +1,4 @@
-package com.mini.buting.api.chat.domain;
+package com.mini.buting.api.chat.domain.payload;
 
 public record TextPayload(String text) implements Payload {
 }
