+++ src/main/java/com/mini/buting/api/chat/domain/payload/VotePayload.java
@@ -1,4 +1,4 @@
-package com.mini.buting.api.chat.domain;
+package com.mini.buting.api.chat.domain.payload;
 
 public record VotePayload(Long voteId, String title) implements Payload {
 }
